Actualizar .gitignore para incluir archivos de migraciones, corregir tipos de datos en el modelo Product y ajustar la cadena de conexión en appsettings.

+++ Models/Product.cs
@@ -5,9 +5,9 @@ namespace TiendaOnline1.Models{
         public string? Id { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public string ImageUrl { get; set; } = string.Empty;
-        public int Stock { get; set; }
+        public double Stock { get; set; }
         public DateTime CreatedAt { get; set; } = DateTime.Now;
     }
 }
